qt5-build.eclass: make sure we enable the QML debugging support
The code comment was incorrect, the flag *is* used, at least in Qt 5.6
and later. It obviously applies only to qtdeclarative, but for some
reason the flag can only be toggled from qtbase's configure... meh.

We don't allow turning this functionality off via a USE flag because
it would be a mess to implement properly, and its meaning would be
rather obscure for most users. It's a small amount of code anyway, and
it doesn't require any additional dependencies.

The configure already enabled -qml-debug by default, so this commit
effectively doesn't change anything in practice.